requirements.txt
rdflib==7.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy>=1.21.0,<2.0.0
jsonschema==4.19.0
pyld==2.0.3
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
lxml==4.9.3
beautifulsoup4==4.12.2
textblob==0.17.1
jieba==0.42.1
scikit-learn==1.3.0
spacy==3.6.1
transformers==4.33.2
torch==2.0.1